# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/socks/{id}":{"put":{"tags":["socks-controller"],"summary":"Обновить информацию о носках","description":"Обновляет информацию о носках по ID.","operationId":"updateSock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSocksDto"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Носок не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Носок успешно обновлён","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SocksDto"}}}}}}},"/api/socks/outcome":{"post":{"tags":["socks-controller"],"summary":"Регистрация отпуска носков","description":"Регистрирует отпуска носков по цвету, проценту хлопка и количеству.","operationId":"registerOutcome","parameters":[{"name":"color","in":"query","required":true,"schema":{"type":"string"}},{"name":"cottonPart","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Неверные данные ввода","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Носки успешно зарегистрированы","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SocksDto"}}}}}}},"/api/socks/income":{"post":{"tags":["socks-controller"],"summary":"Регистрация прихода носков","description":"Регистрирует поступление носков по цвету, проценту хлопка и количеству.","operationId":"registerIncome","parameters":[{"name":"color","in":"query","required":true,"schema":{"type":"string"}},{"name":"cottonPart","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Неверные данные ввода","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Носки успешно зарегистрированы","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SocksDto"}}}}}}},"/api/socks/batch":{"post":{"tags":["socks-controller"],"summary":"Загрузить партию носков","description":"Загружает партию носков из файла формата CSV.","operationId":"uploadBatch","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Неверный формат файла","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Партия успешно обработана","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SocksDto"}}}}}}}},"/api/socks":{"get":{"tags":["socks-controller"],"summary":"Получение общего количества носков с фильтрацией","description":"Получает общее количество носков с возможностью фильтрации по цвету, оператор сравнения (moreThan, lessThan, equal), диапазону процентного содержания хлопка (например, от 30 до 70%).","operationId":"getSocksQuantity","parameters":[{"name":"color","in":"query","required":false,"schema":{"type":"string"}},{"name":"operation","in":"query","required":false,"schema":{"type":"string","enum":["MORE_THAN","LESS_THAN","EQUAL"]}},{"name":"minCottonPart","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"maxCottonPart","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}},{"name":"quantity","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Неверные параметры запроса","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Количество успешно получено","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/socks/all":{"get":{"tags":["socks-controller"],"summary":"Получить все носки","description":"Получает все носки с возможностью фильтрации по цвету, оператор сравнения (moreThan, lessThan, equal), диапазону процентного содержания хлопка (например, от 30 до 70%), сортировкой.","operationId":"getSocks","parameters":[{"name":"color","in":"query","required":false,"schema":{"type":"string"}},{"name":"operation","in":"query","required":false,"schema":{"type":"string","enum":["MORE_THAN","LESS_THAN","EQUAL"]}},{"name":"minCottonPart","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"maxCottonPart","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}},{"name":"quantity","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","enum":["COLOR","COTTON_PART"]}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Неверные параметры запроса","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Носки успешно получены","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SocksDto"}}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"status":{"type":"string"},"reason":{"type":"string"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"UpdateSocksDto":{"type":"object","properties":{"color":{"type":"string"},"cottonPart":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"SocksDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"color":{"type":"string"},"cottonPart":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}}}}}