# src/main/java/com/ppxxd/backspark_test/exception/ErrorResponse.java
package com.ppxxd.backspark_test.exception;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private String status;
    private String reason;
    private String message;
    private String timestamp;
}