# src/main/java/com/ppxxd/backspark_test/socks/dto/SocksDto.java
package com.ppxxd.backspark_test.socks.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class SocksDto {
    private Long id;
    private String color;
    private Integer cottonPart;
    private Integer quantity;
}